requirements.txt
chromadb>=0.4.22
pypdf2>=3.0.1
python-docx>=1.1.0
langchain>=0.1.0
ollama>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
tqdm>=4.66.0
psutil>=5.9.0
pynvml>=11.5.0
newspaper3k>=0.2.8
pyaudio>=0.2.11
keyboard>=0.13.5
pygame>=2.5.0
torch>=2.0.0
duckduckgo-search>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
customtkinter>=5.2.0
Pillow>=10.0.0
cryptography>=41.0.0
trimesh>=3.20.0
open3d>=0.18.0

